Fix race condition returning to orbit
If `remaining_runs` was decremented after `reload_landing()` was called it could cause it to attempt to return to orbit while the game is loading into the landing. This turned into a race condition, as loading into the landing could cause the UI to close during the return to orbit sequence.